Use The Unlicense for public domain dedication
This software is in public domain: it was derived from the public domain
code and includes only public domain code. To indicate this for NPM, we
used "SEE LICENSE IN COPYING.txt" file in package.json (because "public
domain" is not in SPDX list (https://spdx.org/licenses/), which wasn't
ideal, as it required manual review to figure out the licensing
situation.

This commit switches our text for public domain dedication to
The Unlicense. Now automated tools and GitHub can display "Unlicense",
making it clear that the project is in public domain.

Additionally, this commit adds AUTHORS.md file to list all authors of
this code (useful for legal purposes, as in some jurisdictions even
public domain code requires attribution), and a new pull request
template requiring contributors to explicitly release their contribution
to the public domain for non-trivial changes.

File ISSUE_TEMPLATE.md has been removed, as hopefully this change will
stop issues about licensing.